Clean up git aliases and functions
To simplify my workflow, 'origin' is now always my "most local" repo, which (for active projects that I'm not always wanting to make public right away) will have URL git@llamail.com, else a typical github(-like) URL.

In the latter case (a public URL), it'll likely be one of my own forks.

The 'gitnbr' command will always push the new branch it creates to 'origin', which used to be annoying when I wasn't ready to publicize such branches yet it pointed to github. Now it'll point to llamail.com in cases where privacy is important, so it'll "just work".

That seems more sensible than adding the script-fu to divine the remote for the current branch.

For 'gitrmbr', however, as there's an existing branch to delete, it'll (presumably) have an upstream, and that can be fairly easily determined, given that branch name. So that upstream, not necessarily always 'origin', will be where the remote is deleted along with the local.

And as I don't use the "gitf[dou]" or "gitpend" commands anymore, and the nomenclature they use has changed, it's time to say goodbye to them.